pyproject.toml
[project]
name = "mcp-server-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-groq>=0.3.7",
    "mcp-cli>=0.6",
    "mcp-use>=1.3.8",
    "mcp[cli]>=1.12.4",
    "python-dotenv>=1.1.1",
]
